refactor(models): move shared base contracts
